auth-service/requirements.txt
aiosmtplib==4.0.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
bcrypt==4.3.0
beanie==1.30.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
Flask==2.2.2
fqdn==1.5.1
fsspec==2025.3.2
git-filter-repo==2.47.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
motor==3.7.1
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.6
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.9.0
pymongo==4.13.1
PySocks==1.7.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==3.2.1
python-multipart==0.0.20
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
selenium==4.28.1
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.28.0
trio-websocket==0.11.1
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
